% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\docType{data}
\name{cdata}
\alias{cdata}
\title{Cirrhosis death rate data}
\format{An object of class \code{data.frame} with 46 rows and 6 columns.}
\usage{
cdata
}
\description{
A data frame across 46 states trying to predict the death rate. The variables are:
pop: size of the urban population (%)
the number of births to women between 45 to 49 (actually, the reciprocal of that value, times 100)
consumption of wine per capita
consumption of hard liquor per capita
death rate from cirrhosis
}
\keyword{datasets}

