% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\docType{data}
\name{celllibs}
\alias{celllibs}
\title{Cell information (names and clusters) for 300 cells}
\format{An object of class \code{data.frame} with 300 rows and 6 columns.}
\usage{
celllibs
}
\description{
Cell information (names and clusters) for 300 cells
}
\keyword{datasets}

