% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\docType{data}
\name{pollen2014}
\alias{pollen2014}
\title{Gene expression measurements across 300 cells of 8686 genes}
\format{An object of class \code{matrix} with 8686 rows and 300 columns.}
\usage{
pollen2014
}
\description{
Gene expression measurements across 300 cells of 8686 genes
}
\keyword{datasets}

